Module Js.Float

Provides functions for inspecting and manipulating floats

type t = float

val _NaN : t

val isNaN : t -> bool

val isFinite : t -> bool

val isInteger : t -> bool

Returns true if the value is a finite number with no fractional part

val toExponential : ?digits:int -> t -> string

Formats a number in exponential notation.

raises Invalid_argument
if digits is not in range 0-100

val toFixed : ?digits:int -> t -> string

Formats a number with fixed-point notation.

raises Failure
if digits is not in range 0-100

val toPrecision : ?digits:int -> t -> string

Formats a number with the specified number of significant digits.

raises Invalid_argument
if digits is not in range 1-100

val toString : ?radix:int -> t -> string

Converts a number to a string. Optionally specify a radix (2-36).

raises Invalid_argument
if radix is not in range 2-36

val fromString : string -> t